Moved naming conventions to api, removed references in SettingNamingConvention
