+++ src/main/java/me/zeroeightsix/fiber/api/annotation/Settings.java
@@ -2,7 +2,7 @@ package me.zeroeightsix.fiber.api.annotation;
 
 import me.zeroeightsix.fiber.api.annotation.convention.SettingNamingConvention;
 import me.zeroeightsix.fiber.api.tree.ConfigTree;
-import me.zeroeightsix.fiber.impl.annotation.convention.NoNamingConvention;
+import me.zeroeightsix.fiber.api.annotation.convention.NoNamingConvention;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
