+++ src/test/java/me/zeroeightsix/fiber/impl/annotation/convention/SnakeCaseConventionTest.java
@@ -1,6 +1,6 @@
 package me.zeroeightsix.fiber.impl.annotation.convention;
 
-import me.zeroeightsix.fiber.impl.annotation.convention.SnakeCaseConvention;
+import me.zeroeightsix.fiber.api.annotation.convention.SnakeCaseConvention;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
